refactoring wip
